# docs/swagger.yaml
basePath: /
host: localhost:8084
info:
  contact:
    email: support@vhvplatform.com
    name: API Support
    url: http://www.vhvplatform.com/support
  description: Multi-channel Notification Service with Email, SMS, and Webhook support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Notification Service API
  version: "1.0"
paths:
  /api/v1/dlq:
    get:
      consumes:
      - application/json
      description: Get list of failed notifications from DLQ
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of failed notifications
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get failed notifications
      tags:
      - dlq
  /api/v1/dlq/{id}/retry:
    post:
      consumes:
      - application/json
      description: Retry sending a failed notification
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Notification retried
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Notification not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Retry failed notification
      tags:
      - dlq
  /api/v1/notifications:
    get:
      consumes:
      - application/json
      description: Get list of notifications with pagination
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of notifications
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get notifications
      tags:
      - notifications
  /api/v1/notifications/{id}:
    get:
      consumes:
      - application/json
      description: Get details of a specific notification
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Notification details
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Notification not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get notification by ID
      tags:
      - notifications
  /api/v1/notifications/bulk/email:
    post:
      consumes:
      - application/json
      description: Send email notifications to multiple recipients
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Bulk email request
        in: body
        name: bulk
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "202":
          description: Bulk email accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Send bulk email notifications
      tags:
      - notifications
  /api/v1/notifications/email:
    post:
      consumes:
      - application/json
      description: Send an email notification to recipients
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Email request
        in: body
        name: email
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Email sent successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Send email notification
      tags:
      - notifications
  /api/v1/notifications/sms:
    post:
      consumes:
      - application/json
      description: Send an SMS notification
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: SMS request
        in: body
        name: sms
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: SMS sent successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Send SMS notification
      tags:
      - notifications
  /api/v1/notifications/webhook:
    post:
      consumes:
      - application/json
      description: Send a webhook notification to a specified URL
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Webhook request
        in: body
        name: webhook
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Webhook sent successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Send webhook notification
      tags:
      - notifications
  /api/v1/preferences/{user_id}:
    get:
      consumes:
      - application/json
      description: Get notification preferences for a user
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User preferences
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Preferences not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get user notification preferences
      tags:
      - preferences
    put:
      consumes:
      - application/json
      description: Update notification preferences for a user
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Preferences update
        in: body
        name: preferences
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Preferences updated
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update user notification preferences
      tags:
      - preferences
  /api/v1/scheduled:
    get:
      consumes:
      - application/json
      description: Get list of scheduled notifications
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of schedules
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get scheduled notifications
      tags:
      - scheduled
    post:
      consumes:
      - application/json
      description: Create a new scheduled notification
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Schedule request
        in: body
        name: schedule
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Schedule created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Create scheduled notification
      tags:
      - scheduled
  /api/v1/scheduled/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a scheduled notification
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Schedule deleted
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Schedule not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Delete scheduled notification
      tags:
      - scheduled
    put:
      consumes:
      - application/json
      description: Update an existing scheduled notification
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      - description: Schedule update
        in: body
        name: schedule
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Schedule updated
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Schedule not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update scheduled notification
      tags:
      - scheduled
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"